requirements.txt
anthropic>=0.40.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
bcrypt>=4.1.0
openpyxl>=3.1.0
python-docx>=1.1.0
rich>=13.0.0
pyyaml>=6.0
pdfplumber>=0.10.0
flask>=3.0.0
playwright>=1.40.0
streamlit>=1.35.0
pandas>=2.0.0
python-jobspy>=1.1.0
reportlab>=4.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.0
apscheduler>=3.10.0

# Stripe billing (Pro plan checkout + webhook)
stripe>=8.0.0

# Optional: needed for --ollama mode only
openai>=1.0.0
